Function signal::exec_handler::set_command_line

pub fn set_command_line<P, Ai, A, Ek, Ev, E>(program: P, args: A, environ: E) where
    P: ToCString,
    Ai: ToCString,
    A: IntoIterator<Item = Ai>,
    Ek: ToCString,
    Ev: ToCString,
    E: IntoIterator<Item = (Ek, Ev)>, 

Sets command-line and environment to execute when signal happens

If nothing is set current command-line is used.

You may change it freely even after calling set_handler. But only single command-line may be set for all signal handlers used in this module.

Since version v0.3.0 command-line is executed only if pid of a process matches original pid where set_handler was called. I.e. you need to set_handler again for forked process if that is desired.