Module signal::exec_handler

Set a signal handler that executes command-line

This is supposed to work on panics, deadlocks and event many kinds of memory errors. But may leak file descriptors which have no CLOEXEC flag set

This is used as fail-safe signal handler for the supervisors, which are able to re-execute in place and continue to work. Also may be used for configuration reloading signal (e.g. SIGHUP) if appropriate.

Functions

set_command_line
Sets command-line and environment to execute when signal happens

set_handler
Set a handler for multiple signals. If no set_command_line was called before this function the command-line is set from std::env